Workaround and blockers for BZ 1501941

+++ cfme/tests/cloud_infra_common/test_rest_providers.py
@@ -5,8 +5,9 @@ from cfme import test_requirements
 from cfme.cloud.provider.azure import AzureProvider
 from cfme.common.provider import CloudInfraProvider
 from cfme.utils import error, testgen
+from cfme.utils.blockers import BZ
 from cfme.utils.rest import assert_response
-from cfme.utils.version import pick
+from cfme.utils.version import current_version, pick
 from cfme.utils.wait import wait_for
 
 
@@ -22,7 +23,17 @@ def delete_provider(appliance, name):
         return
 
     prov = provs[0]
-    prov.action.delete()
+
+    # workaround for BZ1501941
+    if current_version() >= '5.9' and BZ(1501941, forced_streams=['5.9', 'upstream']).blocks:
+        prov.action.edit(enabled=False)
+        prov.action.delete()
+        prov.wait_not_exists(num_sec=15, silent_failure=True)
+        if prov.exists:
+            prov.action.delete()
+    else:
+        prov.action.delete()
+
     prov.wait_not_exists(num_sec=30)
 
 
@@ -183,6 +194,7 @@ def test_provider_edit(request, provider_rest, appliance):
 
 
 @pytest.mark.tier(1)
+@pytest.mark.meta(blockers=[BZ(1501941, forced_streams=['5.9', 'upstream'])])
 @pytest.mark.parametrize("method", ["post", "delete"], ids=["POST", "DELETE"])
 def test_provider_delete_from_detail(provider_rest, appliance, method):
     """Tests deletion of the provider from detail using REST API.
@@ -204,6 +216,7 @@ def test_provider_delete_from_detail(provider_rest, appliance, method):
 
 
 @pytest.mark.tier(1)
+@pytest.mark.meta(blockers=[BZ(1501941, forced_streams=['5.9', 'upstream'])])
 def test_provider_delete_from_collection(provider_rest, appliance):
     """Tests deletion of the provider from collection using REST API.
